+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Change default of sequential prefetch base to 2.0
 * Added CHANGELOG
 
 ## v0.1 (October 16, 2024)